% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{appendResults}
\alias{appendResults}
\title{Append a number of results to data}
\usage{
appendResults(dat, funs, nmz = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{PSA data}

\item{funs}{a list of functions}

\item{nmz}{an optional vector specifying a subset of the functions in \code{funs} to run}

\item{verbose}{print what's happening (default=\code{TRUE})}
}
\description{
Append a number of results to data
}
\author{
Pete Dodd
}
