% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_points.R
\name{ci_points}
\alias{ci_points}
\alias{midpnt}
\alias{lopnt}
\alias{hipnt}
\title{Confidence interval low, mid and high points}
\usage{
midpnt(x)

lopnt(x)

hipnt(x)
}
\arguments{
\item{x}{Confidence interval string}
}
\value{
numeric
}
\description{
`midpnt()` returns the midpoint from CIs. Utility function
for extracting midpoint from bracketed uncertainty ranges.

`lopnt()` returns the low point from CIs. Utility function
for extracting low point from bracketed uncertainty ranges.

`hipnt()` returns the high point from CIs. Utility function
for extracting high point from bracketed uncertainty ranges.

Expects something like M (Mlo,Mhi) or M (Mlo - Mhi).
}
\author{
Pete Dodd
}
