% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getAB}
\alias{getAB}
\title{Parameters for a beta distribution}
\usage{
getAB(E, V)
}
\arguments{
\item{E}{Mean}

\item{V}{Variance}
}
\value{
\code{list} containing a & b parameter values
}
\description{
Parameters for a beta distribution
}
\author{
Pete Dodd
}
