% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePSA.R
\name{makePSA}
\alias{makePSA}
\title{Facilitates generating a PSA data.table object from a parameter object}
\usage{
makePSA(N, P, dbls = list())
}
\arguments{
\item{N}{number of samples (rows)}

\item{P}{parameter object (eg from parse.parmtable)}

\item{dbls}{list of 2-D named vectors for any parameter pairs generated
jointly (same order as in \code{P})}
}
\value{
\code{data.table} object
}
\description{
Facilitates generating a PSA data.table object from a parameter object
}
\author{
Pete Dodd
}
