% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showParmz.R
\name{makeTestData}
\alias{makeTestData}
\title{Makes a dummy PSA dataset based on a set of parameters for function testing}
\usage{
makeTestData(ncheck, vnames)
}
\arguments{
\item{ncheck}{Number check}

\item{vnames}{Vector of names}
}
\value{
\code{data.frame} of dummy PSA for testing tree functions
}
\description{
Makes a dummy PSA dataset based on a set of parameters for function testing
}
\author{
Pete Dodd
}
