% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTfuns.R
\name{makeTfuns}
\alias{makeTfuns}
\title{Make functions to evaluate tree}
\usage{
makeTfuns(node, qnt = c("cost", "qol"))
}
\arguments{
\item{node}{}

\item{qnt}{}
}
\value{
A \code{list} of functions
}
\description{
Prepares functions that evaluate the strings generated by \code{getAQ}.
}
\details{
The default quantities for mean calculation over tree are `cost` and `qol`,
but others can be specified and will work as long as these variables are defined
at all nodes of the tree.
}
\examples{
tree_fns <- makeTfuns(Dx, qnt=qnts)

dat <-
  data.frame(
    p.screen = 1,
    p.test = 1,
    cfr.tx = 1,
    cfr.notx = 1))
    
tree_fns[[1]](dat)

}
\author{
Pete Dodd
}
