% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{savetreeplot}
\alias{savetreeplot}
\title{Ditch the top of tree on reading}
\usage{
savetreeplot(tree, filename)
}
\arguments{
\item{tree}{tree to plot}

\item{filename}{string}
}
\description{
Ditch the top of tree on reading
}
\author{
Pete Dodd
}
