% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showParmz.R
\name{showAllParmz}
\alias{showAllParmz}
\title{List the parameters used in the model for use in testing}
\usage{
showAllParmz(TREE)
}
\arguments{
\item{TREE}{}
}
\value{
\code{vector} of unique basic parameters
}
\description{
List the parameters used in the model for use in testing
}
\author{
Pete Dodd
}
