% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{top}
\alias{top}
\title{Drop the top of tree on reading}
\usage{
top(x)
}
\arguments{
\item{x}{data.tree object}
}
\value{
data.tree object
}
\description{
Drop the top of tree on reading
}
\author{
Pete Dodd
}
