% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_points.R
\name{ci_points}
\alias{ci_points}
\alias{midpnt}
\alias{lopnt}
\alias{hipnt}
\title{Utility functions for confidence interval low, mid and high points}
\usage{
midpnt(x)

lopnt(x)

hipnt(x)
}
\arguments{
\item{x}{Confidence interval string}
}
\value{
numeric
}
\description{
`midpnt()` returns the midpoint from CIs. 

`lopnt()` returns the low point from CIs.

`hipnt()` returns the high point from CIs.

Extracts from bracketed uncertainty ranges.
Expects something like M (Mlo,Mhi) or M (Mlo - Mhi).
}
\examples{
midpnt("1 (0,2)")
hipnt("1 (0-200)")
lopnt("0 (-1,2)")  ##TODO: error
}
\author{
Pete Dodd
}
